+++ plugins/EnhancedHotSwapEnabler/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [1.1.0] - 2025-08-19
+
 ### Added
 
 - Configurations and settings supports
@@ -19,5 +21,6 @@
 - Not editing parameters for non-debug starts
 - Wrong gradle plugin class name
 
-[Unreleased]: https://github.com/ghostflyby/IntelliJ-Plugins/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/ghostflyby/IntelliJ-Plugins/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/ghostflyby/IntelliJ-Plugins/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/ghostflyby/IntelliJ-Plugins/commits/v1.0.0